Working binder unit without inheritance or activation
